refactor(DEV-10391): replace `<InvoiceActionMenu/>` with a property on `<InvoicesTable/>`
