--- packages/sdk-react/src/components/receivables/InvoicesTable/InvoiceActionMenu/index.ts
@@ -1,1 +1,0 @@
-export * from './InvoiceActionMenu';
