+++ packages/sdk-react/src/components/receivables/InvoicesTable/InvoicesTable.stories.tsx
@@ -21,4 +21,30 @@ export const FullPermissions: Story = {
   ),
 };
 
+export const WithCustomActions: Story = {
+  args: {
+    onRowClick: action('onRowClick'),
+    onRowActionClick: action('onRowActionClick'),
+  },
+  render: (args) => (
+    <div style={{ height: 500 }}>
+      <InvoicesTable
+        {...args}
+        rowActions={{
+          draft: ['edit', 'delete'], // remove 'view' from default
+          issued: [
+            'view',
+            'send',
+            'cancel',
+            'pay',
+            'partiallyPay', // 'copyPaymentLink', 'partiallyPay', 'overduePayment' are not default
+            'copyPaymentLink',
+            'overduePayment',
+          ],
+        }}
+      />
+    </div>
+  ),
+};
+
 export default meta;
